[develop] Edit readme to mount server volume; Set up debug logging and start test example in docker by default;

+++ scripts/develop/config_server.py
@@ -144,8 +144,10 @@ def make():
   example_config["siteUrl"] = "http://" + config.option("siteUrl") + ":8000/"
   example_config["apiUrl"] = "web-apps/apps/api/documents/api.js"
   example_config["preloaderUrl"] = "web-apps/apps/api/documents/cache-scripts.html"
-  json_file = git_dir + "/document-server-integration/web/documentserver-example/nodejs/config/local-development-" + base.host_platform() + ".json"
-  base.writeFile(json_file, json.dumps({"server": example_config}, indent=2))
+  json_dir = git_dir + "/document-server-integration/web/documentserver-example/nodejs/config/"
+  json_file = json_dir + "/local-development-" + base.host_platform() + ".json"
+  if base.is_exist(json_dir):
+    base.writeFile(json_file, json.dumps({"server": example_config}, indent=2))
   
   os.chdir(old_cur)
   return
